backend/requirements.txt
python-dotenv
Flask-SQLAlchemy
Flask
Flask-CORS
psycopg2
flask-socketio
gunicorn
gevent